requirements.txt
pip~=25.2
pillow~=11.3.0
tqdm~=4.67.1
opencv-python~=4.12.0.88
numpy~=2.2.6